[Libomptarget] Add a buffer of shared memory for memory globalizations.
Summary:
Currently the runtime implementation of `__kmpc_alloc_shared` is extremely slow because it allocated memory for each thread individually. This patch adds a small buffer for the threads to share data and will greatly improve performance for builds where all globalization could not be optimized out.

Depends on D97680