[OpenMP] Add OpenMP variant extension to keep the unmangled name
The OpenMP variant clause current ensures unique function identifiers by
mangling the name with `$ompvariant...`. This patch adds an extension to
prevent this by stripping off the mangling during code generation. Doing
this will conflict with the other symbols in the module, but it will
allow us to overload the names of existing symbols if their definitions
are allowed to be dropped. This is most useful for overloading the
meaning of math functions without changing the symbol so they can still
be identified by LLVM.

Differential Revision: https://reviews.llvm.org/D122760