[Libomptarget] Add support for offloading binaries in libomptarget
Summary:
The previous path changed the linker wrapper to embed the offloading
binary format inside the target image instead. This will allow us to
more generically bundle metadata with these images, such as requires
clauses or the target architecture it was compiled for. In order to
parse these classes I pulled out the implementation of the binary from
llvm/Object/OffloadBinary.h and specialized it for this use-case. This
is so we don't need to pull in the LLVM libraries to handle this.

I wasn't sure how to handle this best, so I introduced a new type that
replaces the old __tgt_device_image struct that we can expand inside
the runtime library. I made the new __tgt_device_binary struct pretty
much the same for now. In the future we could change this struct to
pretty much be the OffloadBinary class, but I wasn't sure if I wanted
the plugins to require dependencies more complicated than some pointers.

This approach also requires copying the pointers to make new structs,
but we're referencing the same data buffer so copying a few pointers
into a vector isn't likely to introduce any significant overhead.

Let me know if we should use some other approach to get this done.

Differential Revision: https://reviews.llvm.org/D127432