[libc] Remove subtarget dependant operations in the GPU build
Summary:
Currently we have a few features that depend on what architecture was
build with `-march`. Things like `__has_builtin` or `__CUDA_ARCH` will
key off of this to provide divergent behaviour. In order to make the GPU
`libc` build as a single IR blob we need to make every option handled
equally in the source code and let the backend handle the differences.

This patch simply pulls out the remaining few cases where this still
existed. One noticable regression was the automatic detection of the
AMDGPU frequency (attempted). However it wasn't exactly correct in the
first place, so defaulting to the most recent value is fine.