[libc] Build the GPU libc with unified LLVM-IR
Summary:
The GPU toolchain previously needed to be built multiple times for each
architecture because we had target dependent features that each one
required. A lot of recent changes I've made in clang and LLVM should
make this well-formed enough to be usable here.

The source code has been written in such a way that all target dependent
features are handled by the backend. This allows us to distribute the
source code in LLVM-IR which can be linked into the user's application
and then have the backend run on it with an actual subtarget set. We
copmile without `-march` or `-mcpu` to indicate that we do not want any
`target-cpu` specific metadata to be added. When this finally gets
compiled it will fall back to the subtarget defaults the user has
chosen, which is acceptable because we are currently not exercising any
non-standard per-function behaviour.

This patch keeps most of the existing infrastructure around, primarily
just changing how it is compiled to run the initla tests to verify the
correctness. Future patches will want to completely rework this
interface so that the GPU build is actually two runtime builds simply
targeting NVPTX and AMDGPU individually. Effectively the idea would be
to make the handling the same as x64 Linux / Aarch64 Linux.

This gets rid of `LIBC_GPU_ARCHITECTURES` which used to also enable GPU
build mode so anyone depending on this behaviour will need to change.