[libcxxabi] Enable building the basic 'libcxxabi' for the GPU
Summary:
This patch adds the default config and flags required to get a
functioning `libcxxabi`. Here we mostly borrow the support for the
baremetal target and turn off threads.

Things like `__cxa_guard_acquire` are going to be totally broken for
general use, but might be enough to let us tests things.