[AMDGPU] Use the AMDGPUToolChain when targeting C/C++ directly
Summary:
The getToolChain pass uses the triple to determine which toolchain to
create. Currently the amdgcn-amd-amdhsa triple maps to the
ROCmToolChain which uses things expected to be provided by ROCm.
This is neded for OpenCL, but directly targeting C++ does not want this
since it's primarily being used for creating GPU runtime code. As far as
I know I'm the only user of this, so this shouldn't change anything.

Unfortunately, there's no good logic for detercting this, so I simply
checked ahead of time if the input is either foo.cl or -x cl foo.c
to choose between the two. This allows us to use the AMDGPU target
normally, as otherwise it will error without passing -nogpulib.